Software/requirements.txt
pyserial==3.4
opencv-contrib-python==4.1.1.26
numpy==1.17.4